backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pymongo==4.8.0
motor==3.5.1
email-validator==2.2.0
jinja2==3.1.4
itsdangerous==2.2.0
stripe>=13.0.0
